Save/Rewind: copy state via assignment, small optimizations - Use asignment for struct copy instead of memcpy (ebven though sdcc just calls memcpy for it) - At some point use a smaller, faster memcpy (< 256 bytes) - Disable duplicated zero check for game_state_restore() - Small optimization for checking state_restore_count - 8 bit test since never > 0x99

+++ src/common.h
@@ -206,7 +206,8 @@ typedef struct game_state_data {
 extern game_state_data state;
 extern BCD state_restore_count;
 
-#define STATE_RESTORE_COUNT_GET() (state_restore_count)
+// Can truncate this to uint8 since value should never be over 99 BCD (0x99)
+#define STATE_RESTORE_COUNT_GET() ((uint8_t)state_restore_count)
 
 
 void delay_lowcpu(uint16_t num_frames);
